Add initial Debian 13 adaptation

+++ install/terminal/app-github-cli.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if ! apt-cache policy gh | grep -q 'Candidate: (none)'; then
+    sudo apt install gh -y
+    return 0 2>/dev/null || exit 0
+fi
+
 if [ ! -f /etc/apt/sources.list.d/github-cli.list ]; then
     [ -f /usr/share/keyrings/githubcli-archive-keyring.gpg ] && sudo rm /usr/share/keyrings/githubcli-archive-keyring.gpg
     curl -fsSL https://cli.github.com/packages/githubcli-archive-keyring.gpg | sudo dd of=/usr/share/keyrings/githubcli-archive-keyring.gpg status=none
